Intro Level Semi-Final
Alex if you could give this an art pass - i did quite a bit already so it shouldnt be bad. I also need jenna to implement the checkpoint script to the checkpoint assets i have laid out. Most of the collider boxes are there, but i didnt complete it all